Refactor email option function into a seperate utility file
It doesn't make sense to include the `buildSendEmailOptions` utility
inside the `action-send-email` file, since its not actually used there.

Change-type: patch
